Refactor: 사용자에게 TransactionVo 리스트로 넘겨주던 형식을 TransactionDto 리스트를 넘겨주는 것으로 수정
